// peso_lucky_admin-main/src/pages/Aviator/Record/round/data.d.ts
export interface Params {}
export interface ListItem {
  cancel_operator: string;
  canceled_at: number;
  close_time: number;
  created_at: number;
  end_time: number;
  first_draw_result: string;
  first_draw_time: number;
  first_draw_time_str: string;
  game_code: string;
  game_code_type: string;
  game_name: string;
  game_type: number;
  issue_number: string;
  next_seq: string;
  open_time: number;
  operator: string;
  prize_time: number;
  remark: string;
  round_type: number;
  second_draw_result: string;
  second_draw_time: number;
  second_draw_time_str: string;
  second_settlement_result: string;
  second_settlement_time: number;
  second_settlement_tow_result: string;
  seq: string;
  status: number;
  updated_at: number;
  max_odds: number;
}